requirements.txt
faiss-cpu
groq
langchain-groq
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf
flask>=2.3.0
pandas 
numpy 
scikit-learn 
sentence-transformers 
pinecone-client
python-dotenv
spacy